Replace with using single instance for Enjin provider, current impl causes memory leaks
