+++ AlphaWallet/Core/NFT/Enjin/EnjinOauth.graphql
