+++ AlphaWallet/Core/NFT/Enjin/EnjinToken.graphql
