+++ AlphaWallet/Core/NFT/OpenSea/Types/OpenSeaCollection.swift
