+++ AlphaWallet/Core/NFT/OpenSea/Types/OpenSeaNonFungible.swift
