+++ AlphaWallet/Core/NFT/OpenSea/Types/OpenSeaNonFungibleTokenHandling.swift
