+++ AlphaWallet/Core/NFT/OpenSea/Types/OpenSeaStats.swift
